backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
PyMySQL==1.0.2
python-dotenv==0.19.1
gunicorn==23.0.0
cryptography>=41.0.0
Werkzeug==2.0.3
Flask-CORS==3.0.10